rnacentral/requirements.txt
BeautifulSoup==3.2.1
Django==1.8.13
PyMySQL==0.6.6
PyYAML==3.10
astroid==1.0.1
cx-Oracle==5.1.3
django-cors-headers==0.13
django-debug-toolbar==1.3.0
django-filter==0.9.2
django-maintenancemode==0.10
django-rq==0.7.0
djangorestframework==2.3.12
logilab-common==0.60.1
pycrypto==2.6.1
pylint==1.1.0
python-dateutil==2.1
python-memcached==1.53
requests==2.2.1
selenium==2.39.0
six==1.8.0
sqlparse==0.1.15
versiontools==1.9.1
wsgiref==0.1.2

# markdown support
Markdown==2.3.1
django-markdown-deux==1.0.5
markdown2==2.3.1

# fabric
ecdsa==0.13
Fabric==1.11.1
paramiko==1.17.0

# supervisord
supervisor==3.1.3
meld3==1.0.0

# python-rq
rq==0.5.1
click==3.3
redis==2.10.3

# django compressor
django-appconf==1.0.2
django-compressor==2.0
rcssmin==1.0.6
rjsmin==1.0.12

# django cache machine
django-cache-machine==0.9.1
